Revision before eval CPP05 : ex00 and ex01 done

+++ CPP05/ex00/Bureaucrat.hpp
@@ -8,7 +8,7 @@ class Bureaucrat
 {
 private:
 	std::string const _name;
-	int _grade;
+	size_t _grade;
 	Bureaucrat(void);
 
 public:
